api/requirements.txt
fastapi==0.97.0
uvicorn[standard]==0.22.0
nats-py==2.3.1
redis==4.5.5